+++ demos/bookstore/app/utils/render.ts
@@ -1,9 +1,9 @@
 import type { Remix } from '@remix-run/dom'
 import { renderToStream } from '@remix-run/dom/server'
-import { html } from '@remix-run/fetch-router/response-helpers'
+import { createHtmlResponse } from '@remix-run/response/html'
 
 import { resolveFrame } from './frame.tsx'
 
 export function render(element: Remix.RemixElement, init?: ResponseInit) {
-  return html(renderToStream(element, { resolveFrame }), init)
+  return createHtmlResponse(renderToStream(element, { resolveFrame }), init)
 }